requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Flask==3.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
nodeenv==1.9.1
prisma==0.15.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
StrEnum==0.4.15
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
